% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOGs.R
\name{annotateOGs}
\alias{annotateOGs}
\title{annotateOGs}
\usage{
annotateOGs(GOsfromOGs, selectedOGs, ont = "BP", ignoreLevels = 0:3, cores = 1)
}
\arguments{
\item{GOsfromOGs}{data.frame. Contains two columns, the header of the first should be GOs and the header of the second should be OGs. The first column should be a list of GOs and the second one a list of the OGs to which each GO belongs. No empty fields are allowed.}

\item{selectedOGs}{character. The OGs that will be evaluated. They should be included in the OGs column in the GOsfromOGs data.frame, but elements not found will be ignored and a warning will list them.}

\item{ont}{character. The ontology of the GO terms to be evaluated. It must include one or more of the following terms: "BP", "MF" or "CC" (such as c("BP", "MF", "CC"). Defaults to "BP".}

\item{ignoreLevels}{numeric. Levels of the GOs that will be ignored during the annotation. The descriptions for the GOs with lower levels are more general. If no filtering is desired set to NULL. Defaults to 0:3.}

\item{cores}{numeric. The number of cores to be used. The greater the number of cores the larger the amount of RAM that will be required. It should be a number greater than 0. Defaults to 1.}
}
\value{
A matrix with the annotation for all OGs.
}
\description{
Function used to annotate OGs by identifying the most relevant GO terms associated to each OG. It obtains the three GO terms with the lowest level belonging to each GO and retrieves their description.
}
\examples{

OGs <- c("OG0009248", "OG0009250", "OG0009251", "OG0009252", "OG0009253", "OG0009254", "OG0009258", "OG0009261", "OG0009267", "OG0009268", "OG0009269", "OG0009271")

annotation <- annotateOGs(sampleGOsfromOGs, OGs)

}
