% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOGs.R
\name{clusterOGs}
\alias{clusterOGs}
\title{clusterOGs}
\usage{
clusterOGs(simMat, clusterits = 20000)
}
\arguments{
\item{simMat}{matrix. Similarity matrix generated with the getSimMat function.}

\item{clusterits}{numeric. Maximal number of iterations to be used by apcluster. Defaults to 20000.}
}
\value{
A data.frame with the clusters which can be used to generate the plot.
}
\description{
Function used to cluster the Orthologous Groups (OGs) using the semantic similarity of their Gene Ontology annotations (GOs).
}
\examples{

clusters <- clusterOGs(sampleSimMat)

}
