% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSemSim.R
\name{getSemSim}
\alias{getSemSim}
\title{getSemSim}
\usage{
getSemSim(GOsfromOGs, selectedOGs, ont = "BP", measure = "Wang", cores = 1)
}
\arguments{
\item{GOsfromOGs}{data.frame. Contains two columns, the header of the first should be GOs and the header of the second should be OGs. The first column should be a list of GOs and the second one a list of the OGs to which each GO belongs. No empty fields are allowed.}

\item{selectedOGs}{character. The OGs that will be evaluated. They should be included in the OGs column in the GOsfromOGs data.frame, but elements not found will be ignored and a warning will list them.}

\item{ont}{character. The ontology of the GO terms to be evaluated. It can be either "BP", "MF" or "CC". Defaults to "BP".}

\item{measure}{character. The method used to calculate the similarity. It can be either "Resnik", "Lin", "Rel", "Jiang", "TCSS" or "Wang". Defaults to "Wang".}

\item{cores}{numeric. The number of cores to be used. The greater the number of cores the larger the amount of RAM that will be required. It should be a number greater than 0. Defaults to 1.}
}
\value{
A matrix object with the semantic similarity between all GOs being analyzed.
}
\description{
Function to calculate the semantic similarity between all the Gene Ontology annotations (GOs) in a list using the mgoSim function from the GOSemSim package. The obtained matrix can then be used to obtain the similarity matrix for different sets of Orthologous Groups (OGs), thus to reduce computation times you can use this function to obtain the semantic similarity matrix for the full set of GOs to be evaluated and then use it to evaluate different sets of OGs.
}
\examples{

OGs <- c("OG0009248", "OG0009250", "OG0009251", "OG0009252", "OG0009253", "OG0009254", "OG0009258", "OG0009261", "OG0009267", "OG0009268", "OG0009269", "OG0009271")

semSim <- getSemSim(GOsfromOGs = sampleGOsfromOGs, selectedOGs = OGs)

}
