% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimMat.R
\name{getSimMat}
\alias{getSimMat}
\title{getSimMat}
\usage{
getSimMat(
  GOsfromOGs,
  selectedOGs,
  ont = "BP",
  measure = "Wang",
  combine = "BMA",
  semSim = NULL,
  cores = 1
)
}
\arguments{
\item{GOsfromOGs}{data.frame. Contains two columns, the header of the first should be GOs and the header of the second should be OGs. The first column should be a list of GOs and the second one a list of the OGs to which each GO belongs. No empty fields are allowed.}

\item{selectedOGs}{character. The OGs that will be evaluated. They should be included in the OGs column in the GOsfromOGs data.frame, but elements not found will be ignored and a warning will list them.}

\item{ont}{character. The ontology of the GO terms to be evaluated. It can be either "BP", "MF" or "CC". Defaults to "BP".}

\item{measure}{character. The method used to calculate the similarity. It can be either "Resnik", "Lin", "Rel", "Jiang", "TCSS" or "Wang". Defaults to "Wang".}

\item{combine}{character. The method used to combine similarity scores of multiple GOs. It can be either "max", "avg", "rcmax" or "BMA". It defaults to "BMA".}

\item{semSim}{matrix. Similarity matrix calculated using the getSemSim function. If it has not been calculated it should be set to NULL. Defaults to NULL.}

\item{cores}{numeric. The number of cores to be used. The greater the number of cores the larger the amount of RAM that will be required. It should be a number greater than 0. Defaults to 1.}
}
\value{
A matrix object with the similarities between the OGs being analysed.
}
\description{
Function to calculate the similarity between OGs by combining the semantic similarity scores between the GOs from each pair of OGs being analyzed. If needed, it can use the getSemSim function to calculate the semantic similarity between GOs and then calculate the similarity between OGs, however if you are going to calculate the similarities between different sets of OGs you might want to use the getSemSim function to obtain the full matrix, store it in an object and then use it to calculate the similarity scores between different sets of OGs to reduce times.
}
\examples{

OGs <- c("OG0009248", "OG0009250", "OG0009251", "OG0009252", "OG0009253", "OG0009254", "OG0009258", "OG0009261", "OG0009267", "OG0009268", "OG0009269", "OG0009271")
simMat <- getSimMat(GOsfromOGs = sampleGOsfromOGs, selectedOGs = OGs)

# Using the output of getSemSim
simMat <- getSimMat(GOsfromOGs = sampleGOsfromOGs, selectedOGs = OGs, semSim = sampleSemSim)

}
