% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWordcloud.R
\name{getWordcloud}
\alias{getWordcloud}
\title{getWordcloud}
\usage{
getWordcloud(GOsfromOGs, ont = "BP", clusters, toExclude = NULL)
}
\arguments{
\item{GOsfromOGs}{data.frame. Contains two columns, the header of the first should be GOs and the header of the second should be OGs. The first column should be a list of GOs and the second one a list of the OGs to which each GO belongs. No empty fields are allowed.}

\item{ont}{character. The ontology of the GO terms to be evaluated. It must include one or more of the following terms: "BP", "MF" or "CC" (such as c("BP", "MF", "CC"). Defaults to "BP".}

\item{clusters}{data.frame. Obtained with the clusterOGs function.}

\item{toExclude}{character. Path to the file containing the terms to be excluded. Defaults to NULL.}
}
\value{
data.frame with the wordcloud for each cluster exemplar.
}
\description{
This function is used to get the wordcloud for a list of OGs of interest.
}
\examples{


wcl <- getWordcloud(sampleGOsfromOGs, clusters = sampleClusters, toExclude = NULL)


}
