% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConstellation.R
\name{plotConstellation}
\alias{plotConstellation}
\title{plotConstellation}
\usage{
plotConstellation(
  clusters,
  annotations,
  term = "BP1",
  color = NULL,
  size = NULL
)
}
\arguments{
\item{clusters}{data.frame. Clusters generated with the clusterOGs function.}

\item{annotations}{data.frame. Annotations for each OG. The rownames should be the OGs and it must contain one or more columns with the annotation with their headers.}

\item{term}{character.}

\item{color}{character.}

\item{size}{character.}
}
\value{
A ggplot2 object
}
\description{
Function used to plot the Orthologous Groups (OGs) clustered by their semantic similarity. A polygon comprising all of the OGs in a cluster and the representative member of that cluster is plotted and their functions are displayed.
}
\examples{

annot <- merge(sampleAnnotation, sampleWcl, by.x = 0, by.y = 0)
row.names(annot) <- annot[, 1]
annot <- annot[, -1]
groups <- c("PMAX and LLON", "PMAX and LLON", "PMAX and LLON", "PMAX and LLON", "PMAX and LLON", "PMAX, SMED and LLON", "PMAX, SMED and LLON", "PMAX, SMED and LLON", "PMAX and SMED", "PMAX, SMED and LLON", "PMAX, SMED and LLON", "PMAX and LLON")
size <- c(2, 5, 2, 2, 2, 4, 3, 3, 2, 3, 3, 3)
annot <- data.frame(annot, groups = groups, size = size)
plotConstellation(sampleClusters, annot, term = "BP1", color = "groups", size = "size")

}
