% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGOsfromOGs.R
\name{prepareGOsfromOGs}
\alias{prepareGOsfromOGs}
\title{prepareGOsfromOGs}
\usage{
prepareGOsfromOGs(OGs, GOsdir, method = "EM", species, cores = 1)
}
\arguments{
\item{OGs}{character. The path to the Orthogroups.tsv output of OrthoFinder found in the Orthogroups folder. The first column is expected to be Orthogroup and the rest the species analysed.}

\item{GOsdir}{character. The path to the directory that contains the output of the annotation tool. The file names should be the name of the species followed by ".emapper.annotations" if the method used was eggNOG-mapper (EM) and ".goPredSim.annotations" if method used was goPredSim (GPS).}

\item{method}{character. The methodology used to functionally annotate the genes for each species. EM stands for eggNOG-mapper and GPS for goPredSim. Defaults to EM.}

\item{species}{character. The species to be assessed. They should match the column names in the OrthoFinder output and the file names produced by the functional annotation software.}

\item{cores}{numeric. The number of cores to be used. Defaults to 1.}
}
\value{
A data frame object with two columns, the first with the GOs and the second with the corresponding OGs that contain them.
}
\description{
Function to prepare a data frame linking the Orthologous Groups (OGs) and the Gene Ontology annotations (GOs) associated to them. It requires the results of OrthoFinder to define the OGs and a directory with the functional annotation results from either eggNOG-mapper or goPredSim for each species.
}
\examples{

OGs <- system.file("extdata", "Orthogroups.tsv", package="constellatoR")
GOsdir <- system.file("extdata", package="constellatoR")
species <- c("SMED", "PMAX", "LLON")

GOsfromOGs <- prepareGOsfromOGs(OGs, GOsdir, species = species)

}
